トップに戻るボタンの挙動を変更しました
