+++ db.js
@@ -1,31 +1,11 @@
-// const Pool = require("pg").Pool;
+const Pool = require("pg").Pool;
 
-// const pool = new Pool({
-//     user: "postgres",
-//     host: "localhost",
-//     database: "news",
-//     password: "",
-//     port: 5433,
-// });
+const pool = new Pool({
+    user: "postgres",
+    host: "localhost",
+    database: "news",
+    password: "パスワード",
+    port: 5433,
+});
 
-// module.exports = pool;
-// const Pool = require("pg").Pool;
-
-// const pool = new Pool({
-//     user: "postgres",
-//     host: "localhost",
-//     database: "news",
-//     password: "パスワード",
-//     port: 5433,
-// });
-
-// const pool = new Pool({
-//   user: "postgres",
-//   host: "localhost",
-//   database: "users",
-//   password: "postgres",
-//   port: 5432,
-// });
-
-
-// module.exports = pool;
+module.exports = pool;
