+++ package.json
@@ -24,6 +24,7 @@
     "@types/swiper": "^6.0.0",
     "aos": "^2.3.4",
     "axios": "^1.2.1",
+    "babel-preset-env": "^1.7.0",
     "body-parser": "^1.20.1",
     "cors": "^2.8.5",
     "express": "^4.18.2",
