+++ src/Firebase.js
@@ -1,18 +1,10 @@
-import firebase from 'firebase/app';
+// import firebase from 'firebase/app';
 // import 'firebase/auth';
 import { initializeApp } from "firebase/app";
 import { getFirestore } from "firebase/firestore"
 import { getAuth, GoogleAuthProvider } from 'firebase/auth';
 import { getStorage } from "firebase/storage"
 
-// const firebaseConfig = {
-//   apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-//   authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-//   projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-//   storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-//   messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGE_SENDER_ID,
-//   appId: process.env.REACT_APP_FIREBASE_SENDER_ID,
-// };
 
 const firebaseConfig = {
   apiKey: "AIzaSyA5Fd_9792kFQ4H1FR50WvvoW8khWyv8lk",
